requirements.txt
python-dotenv
streamlit
azure-storage-blob
azure-ai-documentintelligence
azure.core